requirements.txt
streamlit
easyocr
gTTS
pillow
transformers
deep_translator
validators
streamlit-js-eval